Fix for build errors of exporting Props
