+++ packages/react-compass/src/natural-drawer/index.ts
@@ -1,16 +1,11 @@
-import {NaturalDrawerWrapper} from './natural-drawer'
+import NaturalDrawer from './natural-drawer'
 import {NaturalDrawerContent} from './natural-drawer-content'
 import {NaturalDrawerHeader} from './natural-drawer-header'
-
-const NaturalDrawer = NaturalDrawerWrapper as typeof NaturalDrawerWrapper & {
-  Header: typeof NaturalDrawerHeader
-  Content: typeof NaturalDrawerContent
-}
 
 NaturalDrawer.Header = NaturalDrawerHeader
 NaturalDrawer.Content = NaturalDrawerContent
 NaturalDrawer.displayName = 'Drawer'
 
-export type {NaturalDrawerProps} from './types'
+export type {NaturalDrawerProps} from './natural-drawer'
 
 export default NaturalDrawer
