+++ packages/react-compass/src/pagination/index.ts
@@ -1,11 +1,7 @@
 import ItemCounting from './components/ItemCounting'
-import Pagination from './pagination'
+import ComposablePagination from './pagination'
 
 export type {PaginationProps} from './pagination'
-
-const ComposablePagination = Pagination as typeof Pagination & {
-  ItemCounting: typeof ItemCounting
-}
 
 ComposablePagination.displayName = 'Pagination'
 ComposablePagination.ItemCounting = ItemCounting
