+++ packages/react-compass/src/rich-text-editor/rich-text-editor.tsx
@@ -15,8 +15,12 @@ import isEqual from 'lodash/isEqual'
 import React from 'react'
 import CssInjection from '../utils/objectToCss/CssInjection'
 import {useDOMRef} from '../utils/use-dom-ref'
+import * as controls from './controls'
+import Control from './controls/Control'
+import ControlsGroup from './controls/ControlsGroup'
 import {RichTextEditorProvider} from './rich-text-editor.context'
 import styles from './styles/rich-text-editor.module.css'
+import Toolbar from './toolbar'
 
 interface Props {
   children?: React.ReactNode
@@ -36,6 +40,7 @@ interface StorageCount {
 export type RichTextEditorProps = Props &
   Omit<React.HTMLAttributes<HTMLDivElement>, keyof Props>
 
+// eslint-disable-next-line react-refresh/only-export-components
 const RichTextEditor = React.forwardRef<HTMLDivElement, RichTextEditorProps>(
   (props, ref) => {
     const {
@@ -132,4 +137,37 @@ const RichTextEditor = React.forwardRef<HTMLDivElement, RichTextEditorProps>(
     )
   },
 )
-export default RichTextEditor
+export default RichTextEditor as typeof RichTextEditor & {
+  Control: typeof Control
+  ControlsGroup: typeof ControlsGroup
+  Toolbar: typeof Toolbar
+  Bold: typeof controls.BoldControl
+  Italic: typeof controls.ItalicControl
+  Strikethrough: typeof controls.StrikeThroughControl
+  Underline: typeof controls.UnderlineControl
+  H1: typeof controls.H1Control
+  H2: typeof controls.H2Control
+  H3: typeof controls.H3Control
+  H4: typeof controls.H4Control
+  H5: typeof controls.H5Control
+  H6: typeof controls.H6Control
+  BulletList: typeof controls.BulletListControl
+  OrderedList: typeof controls.OrderedListControl
+  Link: typeof controls.LinkControl
+  Unlink: typeof controls.UnlinkControl
+  Image: typeof controls.ImageControl
+  Blockquote: typeof controls.BlockquoteControl
+  AlignLeft: typeof controls.AlignLeftControl
+  AlignRight: typeof controls.AlignRightControl
+  AlignCenter: typeof controls.AlignCenterControl
+  AlignJustify: typeof controls.AlignJustifyControl
+  Superscript: typeof controls.SuperscriptControl
+  Subscript: typeof controls.SubscriptControl
+  CodeBlock: typeof controls.CodeBlockControl
+  ColorControl: typeof controls.ColorControl
+  HeadingsControl: typeof controls.HeadingsControl
+  TextAlginmentSelector: typeof controls.TextAlignmentSelectorControl
+  Hr: typeof controls.HrControl
+  Undo: typeof controls.UndoControl
+  Redo: typeof controls.RedoControl
+}
