+++ packages/react-compass/src/tag-box/index.ts
@@ -1,12 +1,8 @@
-import TagBox from './tag-box'
+import ComposableTagBox from './tag-box'
 import TagBoxAction from './tag-box-action'
 
 export type {TagBoxProps} from './tag-box'
 export type {TagBoxActionProps} from './tag-box-action'
-
-const ComposableTagBox = TagBox as typeof TagBox & {
-  Action: typeof TagBoxAction
-}
 
 ComposableTagBox.Action = TagBoxAction
